requirements.txt
setuptools>=68.0.0
wheel>=0.40.0
python-chess>=1.999
numpy>=1.22.0
moviepy>=1.0.3
Pillow>=10.0.0
svgwrite>=1.4.3 